+++ crates/safe-mmio/src/lib.rs
@@ -16,7 +16,13 @@ mod physical;
 mod volatile_mmio;
 
 use crate::fields::{ReadOnly, ReadPure, ReadPureWrite, ReadWrite, WriteOnly};
-use core::{array, fmt::Debug, marker::PhantomData, ops::Deref, ptr, ptr::NonNull};
+use core::{
+    array,
+    fmt::Debug,
+    marker::PhantomData,
+    ops::{Deref, Range},
+    ptr::{self, NonNull, slice_from_raw_parts_mut},
+};
 pub use physical::PhysicalInstance;
 use zerocopy::{FromBytes, Immutable, IntoBytes};
 
@@ -144,6 +150,27 @@ impl<T: Immutable + IntoBytes> UniqueMmioPointer<'_, ReadWrite<T>> {
     }
 }
 
+impl<T: FromBytes + Immutable + IntoBytes> UniqueMmioPointer<'_, ReadWrite<T>> {
+    /// Performs an MMIO read of the entire `T`, applies the given function to it, and then performs
+    /// an MMIO write of the resulting value.
+    ///
+    /// This is equivalent to calling [`read`](Self::read) then [`write`](Self::write).
+    pub fn modify(&mut self, f: impl FnOnce(T) -> T) {
+        let value = self.read();
+        self.write(f(value));
+    }
+
+    /// Performs an MMIO read of the entire `T`, calls the given function to modify it, and then
+    /// performs an MMIO write of the resulting value.
+    ///
+    /// This is equivalent to calling [`read`](Self::read) then [`write`](Self::write).
+    pub fn modify_mut(&mut self, f: impl FnOnce(&mut T)) {
+        let mut value = self.read();
+        f(&mut value);
+        self.write(value);
+    }
+}
+
 impl<T: Immutable + IntoBytes> UniqueMmioPointer<'_, ReadPureWrite<T>> {
     /// Performs an MMIO write of the entire `T`.
     pub fn write(&mut self, value: T) {
@@ -152,6 +179,27 @@ impl<T: Immutable + IntoBytes> UniqueMmioPointer<'_, ReadPureWrite<T>> {
         unsafe {
             self.write_unsafe(ReadPureWrite(value));
         }
+    }
+}
+
+impl<T: FromBytes + Immutable + IntoBytes> UniqueMmioPointer<'_, ReadPureWrite<T>> {
+    /// Performs an MMIO read of the entire `T`, applies the given function to it, and then performs
+    /// an MMIO write of the resulting value.
+    ///
+    /// This is equivalent to calling [`read`](Self::read) then [`write`](Self::write).
+    pub fn modify(&mut self, f: impl FnOnce(T) -> T) {
+        let value = self.read();
+        self.write(f(value));
+    }
+
+    /// Performs an MMIO read of the entire `T`, calls the given function to modify it, and then
+    /// performs an MMIO write of the resulting value.
+    ///
+    /// This is equivalent to calling [`read`](Self::read) then [`write`](Self::write).
+    pub fn modify_mut(&mut self, f: impl FnOnce(&mut T)) {
+        let mut value = self.read();
+        f(&mut value);
+        self.write(value);
     }
 }
 
@@ -202,6 +250,55 @@ impl<'a, T> UniqueMmioPointer<'a, [T]> {
         Some(unsafe { self.child(regs) })
     }
 
+    /// Returns a `UniqueMmioPointer` to a range of elements of this slice, or `None` if the range
+    /// is out of bounds.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use safe_mmio::{UniqueMmioPointer, fields::ReadWrite};
+    ///
+    /// let mut slice: UniqueMmioPointer<[ReadWrite<u32>]>;
+    /// # let mut fake = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+    /// # slice = UniqueMmioPointer::from(fake.as_mut_slice());
+    /// let mut range = slice.get_range(1..3).unwrap();
+    /// range.get(0).unwrap().write(100);
+    /// range.get(1).unwrap().write(200);
+    /// assert_eq!(None, range.get(2));
+    /// assert_eq!(100, slice.get(1).unwrap().read());
+    /// assert_eq!(200, slice.get(2).unwrap().read());
+    /// ```
+    pub fn get_range(&mut self, range: Range<usize>) -> Option<UniqueMmioPointer<'_, [T]>> {
+        if range.start > range.end || range.end > self.0.len() {
+            return None;
+        }
+
+        let regs_start = if !range.is_empty() {
+            // SAFETY: self.ptr_mut() is guaranteed to return a pointer that is valid for MMIO and
+            // unique, as promised by the caller of `UniqueMmioPointer::new`. range.start is within the
+            // boundaries of the slice.
+            unsafe { &raw mut (*self.ptr_mut())[range.start] }
+        } else {
+            // Based on the documentation of core::slice::from_raw_parts_mut, NonNull::dangling()
+            // should be used for creating zero-length slices.
+            NonNull::dangling().as_ptr()
+        };
+
+        let regs = NonNull::new(slice_from_raw_parts_mut(regs_start, range.len())).unwrap();
+
+        // SAFETY: We created regs from the valid start address of regs_start and `range` is within
+        // the boundaries of self.regs, so it must also be valid, unique and within the allocation
+        // of self.regs.
+        Some(unsafe { self.child(regs) })
+    }
+
+    /// Returns a new iterator of the items of the slice.
+    pub fn iter(&mut self) -> UniqueMmioPointerIterator<'_, T> {
+        UniqueMmioPointerIterator {
+            tail: self.reborrow(),
+        }
+    }
+
     /// Returns a `UniqueMmioPointer` to an element of this slice, or `None` if the index is out of
     /// bounds.
     ///
@@ -237,7 +334,7 @@ impl<'a, T> UniqueMmioPointer<'a, [T]> {
 
 impl<'a, T, const LEN: usize> UniqueMmioPointer<'a, [T; LEN]> {
     /// Splits a `UniqueMmioPointer` to an array into an array of `UniqueMmioPointer`s.
-    pub fn split(&mut self) -> [UniqueMmioPointer<'_, T>; LEN] {
+    pub fn split(mut self) -> [UniqueMmioPointer<'a, T>; LEN] {
         array::from_fn(|i| {
             UniqueMmioPointer(SharedMmioPointer {
                 // SAFETY: self.regs is always unique and valid for MMIO access. We make sure the
@@ -305,6 +402,55 @@ impl<'a, T, const LEN: usize> UniqueMmioPointer<'a, [T; LEN]> {
         // SAFETY: We created regs from the raw array in self.regs, so it must also be valid, unique
         // and within the allocation of self.regs.
         Some(unsafe { self.child(regs) })
+    }
+
+    /// Returns a `UniqueMmioPointer` to a range of elements of this array, or `None` if the range
+    /// is out of bounds.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use safe_mmio::{UniqueMmioPointer, fields::ReadWrite};
+    ///
+    /// let mut slice: UniqueMmioPointer<[ReadWrite<u32>; 3]>;
+    /// # let mut fake = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+    /// # slice = UniqueMmioPointer::from(&mut fake);
+    /// let mut range = slice.get_range(1..3).unwrap();
+    /// range.get(0).unwrap().write(100);
+    /// range.get(1).unwrap().write(200);
+    /// assert_eq!(None, range.get(2));
+    /// assert_eq!(100, slice.get(1).unwrap().read());
+    /// assert_eq!(200, slice.get(2).unwrap().read());
+    /// ```
+    pub fn get_range(&mut self, range: Range<usize>) -> Option<UniqueMmioPointer<'_, [T]>> {
+        if range.start > range.end || range.end > LEN {
+            return None;
+        }
+
+        let regs_start = if !range.is_empty() {
+            // SAFETY: self.ptr_mut() is guaranteed to return a pointer that is valid for MMIO and
+            // unique, as promised by the caller of `UniqueMmioPointer::new`. range.start is within the
+            // boundaries of the array.
+            unsafe { &raw mut (*self.ptr_mut())[range.start] }
+        } else {
+            // Based on the documentation of core::slice::from_raw_parts_mut, NonNull::dangling()
+            // should be used for creating zero-length slices.
+            NonNull::dangling().as_ptr()
+        };
+
+        let regs = NonNull::new(slice_from_raw_parts_mut(regs_start, range.len())).unwrap();
+
+        // SAFETY: We created regs from the valid start address of regs_start and `range` is within
+        // the boundaries of self.regs, so it must also be valid, unique and within the allocation
+        // of self.regs.
+        Some(unsafe { self.child(regs) })
+    }
+
+    /// Returns a new iterator to the items of the array.
+    pub fn iter(&mut self) -> UniqueMmioPointerIterator<'_, T> {
+        UniqueMmioPointerIterator {
+            tail: self.as_mut_slice(),
+        }
     }
 
     /// Returns a `UniqueMmioPointer` to an element of this array, or `None` if the index is out of
@@ -395,6 +541,72 @@ impl<'a, T: ?Sized> Deref for UniqueMmioPointer<'a, T> {
 
     fn deref(&self) -> &Self::Target {
         &self.0
+    }
+}
+
+impl<'a, T> IntoIterator for UniqueMmioPointer<'a, [T]> {
+    type Item = UniqueMmioPointer<'a, T>;
+
+    type IntoIter = UniqueMmioPointerIterator<'a, T>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        UniqueMmioPointerIterator { tail: self }
+    }
+}
+
+impl<'a, T, const LEN: usize> IntoIterator for UniqueMmioPointer<'a, [T; LEN]> {
+    type Item = UniqueMmioPointer<'a, T>;
+
+    type IntoIter = UniqueMmioPointerIterator<'a, T>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        UniqueMmioPointerIterator { tail: self.into() }
+    }
+}
+
+/// Iterator over a `UniqueMmioPointer` slice, yielding pointers to items.
+///
+/// This iterator advances by splitting off the head element and shortening the
+/// remaining tail.
+#[derive(Debug)]
+pub struct UniqueMmioPointerIterator<'a, T> {
+    tail: UniqueMmioPointer<'a, [T]>,
+}
+
+impl<'a, T> Iterator for UniqueMmioPointerIterator<'a, T> {
+    type Item = UniqueMmioPointer<'a, T>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if !self.tail.is_empty() {
+            // SAFETY: self.ptr_mut() is guaranteed to return a pointer that is valid for MMIO and
+            // unique, as promised by the caller of `UniqueMmioPointer::new` and the slice is
+            // not empty.
+            let regs_head = NonNull::new(unsafe { &raw mut (*self.tail.ptr_mut())[0] }).unwrap();
+
+            // SAFETY: regs_head is created from self.tail so it is valid and within the range of
+            // the original pointer. There no other further split_child calls to the same child and
+            // self.tail is moved by one in the following lines.
+            let head = unsafe { self.tail.split_child(regs_head) };
+
+            let regs_tail = NonNull::new(slice_from_raw_parts_mut(
+                regs_head.as_ptr().wrapping_add(1),
+                self.tail.len() - 1,
+            ))
+            .unwrap();
+
+            // SAFETY: regs is created from self.tail so it is valid and within the range of the
+            // original pointer. The new pointer overwrites the original so it cannot be used
+            // afterwards, and there are no further calls to split_child().
+            self.tail = unsafe { self.tail.split_child(regs_tail) };
+
+            Some(head)
+        } else {
+            None
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.tail.len(), Some(self.tail.len()))
     }
 }
 
@@ -531,6 +743,37 @@ impl<'a, T> SharedMmioPointer<'a, [T]> {
         Some(unsafe { self.child(regs) })
     }
 
+    /// Returns a `SharedMmioPointer` to a range of elements of this slice, or `None` if the range
+    /// is out of bounds.
+    pub fn get_range(&self, range: Range<usize>) -> Option<SharedMmioPointer<'_, [T]>> {
+        if range.start > range.end || range.end > self.len() {
+            return None;
+        }
+
+        let regs_start = if !range.is_empty() {
+            // SAFETY: self.ptr_mut() is guaranteed to return a pointer that is valid for MMIO and
+            // unique, as promised by the caller of `UniqueMmioPointer::new`. range.start is within the
+            // boundaries of the slice.
+            unsafe { &raw mut (*self.regs.as_ptr())[range.start] }
+        } else {
+            // Based on the documentation of core::slice::from_raw_parts_mut, NonNull::dangling()
+            // should be used for creating zero-length slices.
+            NonNull::dangling().as_ptr()
+        };
+
+        let regs = NonNull::new(slice_from_raw_parts_mut(regs_start, range.len())).unwrap();
+
+        // SAFETY: We created regs from the valid start address of regs_start and `range` is within
+        // the boundaries of self.regs, so it must also be valid, unique and within the allocation
+        // of self.regs.
+        Some(unsafe { self.child(regs) })
+    }
+
+    /// Returns a new iterator of the items of the slice.
+    pub fn iter(&self) -> SharedMmioPointerIterator<'_, T> {
+        SharedMmioPointerIterator { tail: *self }
+    }
+
     /// Returns the length of the slice.
     pub const fn len(&self) -> usize {
         self.regs.len()
@@ -544,7 +787,7 @@ impl<'a, T> SharedMmioPointer<'a, [T]> {
 
 impl<'a, T, const LEN: usize> SharedMmioPointer<'a, [T; LEN]> {
     /// Splits a `SharedMmioPointer` to an array into an array of `SharedMmioPointer`s.
-    pub fn split(&self) -> [SharedMmioPointer<'a, T>; LEN] {
+    pub fn split(self) -> [SharedMmioPointer<'a, T>; LEN] {
         array::from_fn(|i| SharedMmioPointer {
             // SAFETY: self.regs is always unique and valid for MMIO access. We make sure the
             // pointers we split it into don't overlap, so the same applies to each of them.
@@ -572,6 +815,38 @@ impl<'a, T, const LEN: usize> SharedMmioPointer<'a, [T; LEN]> {
         // SAFETY: We created regs from the raw array in self.regs, so it must also be valid, unique
         // and within the allocation of self.regs.
         Some(unsafe { self.child(regs) })
+    }
+
+    /// Returns a `SharedMmioPointer` to a range of elements of this array, or `None` if the range
+    /// is out of bounds.
+    pub fn get_range(&self, range: Range<usize>) -> Option<SharedMmioPointer<'_, [T]>> {
+        if range.start > range.end || range.end > LEN {
+            return None;
+        }
+
+        let regs_start = if !range.is_empty() {
+            // SAFETY: self.regs is always unique and valid for MMIO access. range.start is within the
+            // boundaries of the slice.
+            unsafe { &raw mut (*self.regs.as_ptr())[range.start] }
+        } else {
+            // Based on the documentation of core::slice::from_raw_parts_mut, NonNull::dangling()
+            // should be used for creating zero-length slices.
+            NonNull::dangling().as_ptr()
+        };
+
+        let regs = NonNull::new(slice_from_raw_parts_mut(regs_start, range.len())).unwrap();
+
+        // SAFETY: We created regs from the valid start address of regs_start and `range` is within
+        // the boundaries of self.regs, so it must also be valid, unique and within the allocation
+        // of self.regs.
+        Some(unsafe { self.child(regs) })
+    }
+
+    /// Returns a new iterator of the items of the array.
+    pub fn iter(&self) -> SharedMmioPointerIterator<'_, T> {
+        SharedMmioPointerIterator {
+            tail: self.as_slice(),
+        }
     }
 }
 
@@ -606,19 +881,86 @@ impl<'a, T> From<SharedMmioPointer<'a, T>> for SharedMmioPointer<'a, [T]> {
     }
 }
 
+impl<'a, T> IntoIterator for SharedMmioPointer<'a, [T]> {
+    type Item = SharedMmioPointer<'a, T>;
+
+    type IntoIter = SharedMmioPointerIterator<'a, T>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        SharedMmioPointerIterator { tail: self }
+    }
+}
+
+impl<'a, T, const LEN: usize> IntoIterator for SharedMmioPointer<'a, [T; LEN]> {
+    type Item = SharedMmioPointer<'a, T>;
+
+    type IntoIter = SharedMmioPointerIterator<'a, T>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        SharedMmioPointerIterator { tail: self.into() }
+    }
+}
+
+/// Iterator over a `SharedMmioPointer` slice, yielding pointers to items.
+///
+/// This iterator advances by creating a head pointer and shortening the
+/// remaining tail.
+#[derive(Clone, Copy, Debug)]
+pub struct SharedMmioPointerIterator<'a, T> {
+    tail: SharedMmioPointer<'a, [T]>,
+}
+
+impl<'a, T> Iterator for SharedMmioPointerIterator<'a, T> {
+    type Item = SharedMmioPointer<'a, T>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if !self.tail.is_empty() {
+            // SAFETY: self.ptr_mut() is guaranteed to return a pointer that is valid for MMIO and
+            // unique, as promised by the caller of `UniqueMmioPointer::new` and the slice is
+            // not empty.
+            let regs_head =
+                NonNull::new(unsafe { &raw mut (*self.tail.regs.as_ptr())[0] }).unwrap();
+
+            // SAFETY: regs_head is created from self.tail so it is valid and within the range of
+            // the original pointer.
+            let head = unsafe { self.tail.child(regs_head) };
+
+            let regs_tail = NonNull::new(slice_from_raw_parts_mut(
+                regs_head.as_ptr().wrapping_add(1),
+                self.tail.len() - 1,
+            ))
+            .unwrap();
+
+            // SAFETY: We created regs from the raw array in self.regs, so it must also be valid,
+            // unique and within the allocation of self.regs.
+            self.tail = unsafe { self.tail.child(regs_tail) };
+
+            Some(head)
+        } else {
+            None
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.tail.len(), Some(self.tail.len()))
+    }
+}
+
 /// Gets a `UniqueMmioPointer` to a field of a type wrapped in a `UniqueMmioPointer`.
 #[macro_export]
 macro_rules! field {
     ($mmio_pointer:expr, $field:ident) => {{
-        // Make sure $mmio_pointer is the right type.
-        let mmio_pointer: &mut $crate::UniqueMmioPointer<_> = &mut $mmio_pointer;
+        _ = &mut $mmio_pointer;
+
         // SAFETY: ptr_mut is guaranteed to return a valid pointer for MMIO, so the pointer to the
-        // field must also be valid. MmioPointer::child gives it the same lifetime as the original
-        // pointer.
+        // field must also be valid. UniqueMmioPointer::child gives it the same lifetime as the
+        // original pointer.
         unsafe {
-            let child_pointer =
-                core::ptr::NonNull::new(&raw mut (*mmio_pointer.ptr_mut()).$field).unwrap();
-            mmio_pointer.child(child_pointer)
+            let child_pointer = core::ptr::NonNull::new(
+                &raw mut (*$crate::UniqueMmioPointer::ptr_mut(&mut $mmio_pointer)).$field,
+            )
+            .unwrap();
+            $crate::UniqueMmioPointer::child(&mut $mmio_pointer, child_pointer)
         }
     }};
 }
@@ -654,16 +996,18 @@ macro_rules! split_fields {
 #[macro_export]
 macro_rules! field_shared {
     ($mmio_pointer:expr, $field:ident) => {{
-        // Make sure $mmio_pointer is the right type.
-        let mmio_pointer: &$crate::SharedMmioPointer<_> = &$mmio_pointer;
+        _ = &$mmio_pointer;
+
         // SAFETY: ptr_mut is guaranteed to return a valid pointer for MMIO, so the pointer to the
         // field must also be valid. MmioPointer::child gives it the same lifetime as the original
         // pointer.
+        #[allow(unused_unsafe, reason = "May be nested")]
         unsafe {
-            let child_pointer =
-                core::ptr::NonNull::new((&raw const (*mmio_pointer.ptr()).$field).cast_mut())
-                    .unwrap();
-            mmio_pointer.child(child_pointer)
+            let child_pointer = core::ptr::NonNull::new(
+                (&raw const (*$crate::SharedMmioPointer::ptr(&$mmio_pointer)).$field).cast_mut(),
+            )
+            .unwrap();
+            $crate::SharedMmioPointer::child(&$mmio_pointer, child_pointer)
         }
     }};
 }
@@ -782,7 +1126,7 @@ mod tests {
         let mut foo = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
         let mut owned = UniqueMmioPointer::from(&mut foo);
 
-        let mut parts = owned.split();
+        let mut parts = owned.reborrow().split();
         assert_eq!(parts[0].read(), 1);
         assert_eq!(parts[1].read(), 2);
         assert_eq!(owned.split()[2].read(), 3);
@@ -870,7 +1214,7 @@ mod tests {
         }
 
         impl Regs {
-            fn from_slice<'a>(slice: &'a mut [ReadPureWrite<u32>]) -> &'a mut Self {
+            fn from_slice(slice: &mut [ReadPureWrite<u32>]) -> &mut Self {
                 let regs_ptr: *mut Self = slice as *mut [ReadPureWrite<u32>] as *mut Self;
                 // SAFETY: `Regs` is repr(transparent) so a reference to its field has the same
                 // metadata as a reference to `Regs``.
@@ -908,8 +1252,8 @@ mod tests {
         assert_eq!(first.read(), 1);
         assert_eq!(second.read(), 2);
 
-        drop(first);
-        drop(second);
+        assert_eq!(first.read(), 1);
+        assert_eq!(second.read(), 2);
 
         assert_eq!(field!(owned, first).read(), 1);
     }
@@ -926,5 +1270,256 @@ mod tests {
 
         assert_eq!(parts[0].read(), 1);
         assert_eq!(parts[1].read(), 2);
+    }
+
+    #[test]
+    fn subfield() {
+        #[repr(C)]
+        struct Regs {
+            subregs: Subregs,
+        }
+
+        #[repr(C)]
+        struct Subregs {
+            field: ReadPureWrite<u32>,
+        }
+
+        let mut foo = Regs {
+            subregs: Subregs {
+                field: ReadPureWrite(0),
+            },
+        };
+        let mut owned: UniqueMmioPointer<Regs> = UniqueMmioPointer::from(&mut foo);
+
+        assert_eq!(
+            field_shared!(field_shared!(owned, subregs), field).read(),
+            0
+        );
+
+        let mut sub = field!(owned, subregs);
+        let mut field = field!(sub, field);
+        field.write(42);
+
+        assert_eq!(foo.subregs.field.0, 42);
+    }
+
+    #[test]
+    fn get_range_slice() {
+        let mut regs = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+
+        {
+            let mut ptr = UniqueMmioPointer::from(&mut regs);
+            let mut slice = ptr.as_mut_slice();
+
+            let range = slice.get_range(100..200);
+            assert!(range.is_none());
+
+            let range = slice.get_range(0..3).unwrap();
+            assert_eq!(range.len(), 3);
+
+            let range = slice.get_range(1..3).unwrap();
+            assert_eq!(range.len(), 2);
+
+            let range = slice.get_range(0..0).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = slice.get_range(2..2).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = slice.get_range(3..3).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = slice.get_range(4..4);
+            assert!(range.is_none());
+
+            let mut range = slice.get_range(3..3).unwrap();
+            let nested_range = range.get_range(0..0).unwrap();
+            assert_eq!(nested_range.len(), 0);
+
+            let nested_range = range.get_range(1..1);
+            assert!(nested_range.is_none());
+        }
+    }
+
+    #[test]
+    fn get_range_array() {
+        let mut regs = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+
+        {
+            let mut ptr = UniqueMmioPointer::from(&mut regs);
+
+            let range = ptr.get_range(100..200);
+            assert!(range.is_none());
+
+            let range = ptr.get_range(0..3).unwrap();
+            assert_eq!(range.len(), 3);
+
+            let range = ptr.get_range(1..3).unwrap();
+            assert_eq!(range.len(), 2);
+
+            let range = ptr.get_range(0..0).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = ptr.get_range(2..2).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = ptr.get_range(3..3).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = ptr.get_range(4..4);
+            assert!(range.is_none());
+
+            let mut range = ptr.get_range(3..3).unwrap();
+            let nested_range = range.get_range(0..0).unwrap();
+            assert_eq!(nested_range.len(), 0);
+
+            let nested_range = range.get_range(1..1);
+            assert!(nested_range.is_none());
+        }
+    }
+
+    #[test]
+    fn shared_get_range_slice() {
+        let regs = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+
+        {
+            let ptr = SharedMmioPointer::from(&regs);
+            let slice = ptr.as_slice();
+
+            let range = slice.get_range(100..200);
+            assert!(range.is_none());
+
+            let range = slice.get_range(0..3).unwrap();
+            assert_eq!(range.len(), 3);
+
+            let range = slice.get_range(1..3).unwrap();
+            assert_eq!(range.len(), 2);
+
+            let range = slice.get_range(0..0).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = slice.get_range(2..2).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = slice.get_range(3..3).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = slice.get_range(4..4);
+            assert!(range.is_none());
+
+            let range = slice.get_range(3..3).unwrap();
+            let nested_range = range.get_range(0..0).unwrap();
+            assert_eq!(nested_range.len(), 0);
+
+            let nested_range = range.get_range(1..1);
+            assert!(nested_range.is_none());
+        }
+    }
+
+    #[test]
+    fn shared_get_range_array() {
+        let regs = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+
+        {
+            let ptr = SharedMmioPointer::from(&regs);
+
+            let range = ptr.get_range(100..200);
+            assert!(range.is_none());
+
+            let range = ptr.get_range(0..3).unwrap();
+            assert_eq!(range.len(), 3);
+
+            let range = ptr.get_range(1..3).unwrap();
+            assert_eq!(range.len(), 2);
+
+            let range = ptr.get_range(0..0).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = ptr.get_range(2..2).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = ptr.get_range(3..3).unwrap();
+            assert_eq!(range.len(), 0);
+
+            let range = ptr.get_range(4..4);
+            assert!(range.is_none());
+
+            let range = ptr.get_range(3..3).unwrap();
+            let nested_range = range.get_range(0..0).unwrap();
+            assert_eq!(nested_range.len(), 0);
+
+            let nested_range = range.get_range(1..1);
+            assert!(nested_range.is_none());
+        }
+    }
+
+    #[test]
+    fn iterator_slice() {
+        let mut regs = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+
+        {
+            let mut ptr = UniqueMmioPointer::from(&mut regs);
+            let mut slice = ptr.as_mut_slice();
+
+            let mut iter = slice.iter();
+
+            iter.next().unwrap().write(4);
+            iter.next().unwrap().write(5);
+            iter.next().unwrap().write(6);
+            assert_eq!(iter.next(), None);
+        }
+
+        assert_eq!(regs[0].0, 4);
+        assert_eq!(regs[1].0, 5);
+        assert_eq!(regs[2].0, 6);
+    }
+
+    #[test]
+    fn iterator_array() {
+        let mut regs = [ReadWrite(1), ReadWrite(2), ReadWrite(3)];
+
+        {
+            let mut ptr = UniqueMmioPointer::from(&mut regs);
+
+            let mut iter = ptr.iter();
+
+            iter.next().unwrap().write(4);
+            iter.next().unwrap().write(5);
+            iter.next().unwrap().write(6);
+            assert_eq!(iter.next(), None);
+        }
+
+        assert_eq!(regs[0].0, 4);
+        assert_eq!(regs[1].0, 5);
+        assert_eq!(regs[2].0, 6);
+    }
+
+    #[test]
+    fn shared_iterator_slice() {
+        let regs = [ReadPureWrite(1), ReadPureWrite(2), ReadPureWrite(3)];
+
+        let ptr = SharedMmioPointer::from(&regs);
+        let slice = ptr.as_slice();
+
+        let mut iter = slice.iter();
+
+        assert_eq!(iter.next().unwrap().read(), 1);
+        assert_eq!(iter.next().unwrap().read(), 2);
+        assert_eq!(iter.next().unwrap().read(), 3);
+        assert_eq!(iter.next(), None);
+    }
+
+    #[test]
+    fn shared_iterator_array() {
+        let regs = [ReadPureWrite(1), ReadPureWrite(2), ReadPureWrite(3)];
+
+        let ptr = SharedMmioPointer::from(&regs);
+
+        let mut iter = ptr.iter();
+
+        assert_eq!(iter.next().unwrap().read(), 1);
+        assert_eq!(iter.next().unwrap().read(), 2);
+        assert_eq!(iter.next().unwrap().read(), 3);
+        assert_eq!(iter.next(), None);
     }
 }